Fix read/write Transit
Previously we were writing all Transit values in one giant pass
(`transit/write writer data`). This can blow memory if data is large,
and is not the intended purpose. Concretely, this changes the data
representation for a JSON-encoded file like:

```
["~#list",[0,1,2,3,4,5,6,7,8,9]]%
```

to:

```
["~#'",0"'"] ["~#'",1"'"] ["~#'",2"'"] ["~#'",3"'"] ...
```

By changing `write-transit` to
incrementally write values from coll and changing `read-transit` to
properly return a lazy seq of values, we avoid this problem. This also
restricts the signature of `write-transit` to strictly accept a coll of
values; this may be more convenient as `x-or-xs` and updated in a future
pass if necessary.