Add support for move-and-replace
Prior to this commit we had `std.io/move`, which only allowed for
move-and-throw-if-already-present.  This commit adds support for
move-and-replace-if-already-present, which allows us to gracefully
handle some of the exceptions we're seeing (multiple calls issued to
moved to the same destination, presumably by multiple racing threads,
and all but the first threads throwing).