Add std.math/variance and std.math/std-dev
This adds functions for computing the variance and standard deviation of
a collection of values, respectively. By default, these functions return
unbiased estimates, using a divisor of `n - 1` instead of `n` (Bessel's
Correction [0]). This is parameterizable by the optional `ddof`
parameter ("delta degreees of freedom") as in NumPy [1], where the divisor is `n -
ddof`; ddof = 0 will result in a biased estimator.

0: https://en.wikipedia.org/wiki/Bessel%27s_correction

1:
https://docs.scipy.org/doc/numpy/reference/generated/numpy.var.html#numpy.var